+++ scripts/dist-win.sh
@@ -2,6 +2,20 @@
 
 # TODO- if I can pass a custom artifact name on the command line I could get this to work.
 
+if [ "${CSC_KEY_PASSWORD}" == "" ]; then
+    echo "CSC_KEY_PASSWORD not set" > /dev/stderr
+    exit 1
+fi
+
+export CSC_LINK=/root/windows-csc/spinn3r.p12
+
+WINDOWS_CSC_DIR=../polar-bookshelf-secrets/windows-csc
+
+# Error: Cannot extract publisher name from code signing certificate, please
+# file issue. As workaround, set win.publisherName: Error: Exit code: 1. Command
+# failed: openssl pkcs12 -nokeys -nodes -passin pass: -nomacver -clcerts -in
+# /root/windows-csc/00C8406FA14CAD991724834F1B0D25C4D3.crt
+
 build_for_arch() {
     arch=${1}
 
@@ -13,9 +27,10 @@ build_for_arch() {
        -v ${PWD##*/}-node-modules:/project/node_modules \
        -v ~/.cache/electron:/root/.cache/electron \
        -v ~/.cache/electron-builder:/root/.cache/electron-builder \
+       -v ${WINDOWS_CSC_DIR}:/root/windows-csc \
        electronuserland/builder:wine bash -c 'yarn && ./node_modules/.bin/electron-builder --config=electron-builder.yml --config.nsis.artifactName=\${name}-\${version}-'${arch}'.\${ext} --'${arch}' --win --publish always'
-
 }
+
 
 build_for_arch x64
 build_for_arch ia32